Formalize Shared Viewer application
